feat: ADR-005 GameCharacterControllerV4 Port 추출 및 이관
Port 인터페이스 (module-core):
- ExpectationV4Port: 기대값 계산 및 GZIP 응답
- PopularCharacterTrackerPort: recordAccess 추가

Adapter 구현체 (module-app):
- ExpectationV4PortAdapter → EquipmentExpectationServiceV4 위임
- PopularCharacterTrackerPortAdapter → PopularCharacterTracker 위임

Controller 이관 (module-web):
- GameCharacterControllerV4 → Port 사용
- Kotlin data class 생성자 호출로 변경
